fix/주석수정
